Separate test scenarios and test runners (for in-memory env)
